requirements.txt
openai
langchain-openai
streamlit
numpy
python-docx
docx2txt
moviepy
